shader2 can now be conditionally inverted
just pushing what i have before starting with textures

+++ LearnOpenGL/src/Shape.h
@@ -63,7 +63,7 @@ public:
 
 	void draw(GLenum mode) {
 		glBindVertexArray(m_VAO);
-		glDrawElements(mode, m_indices.size() * sizeof(unsigned int), GL_UNSIGNED_INT, 0);
+		glDrawElements(mode, (int)(m_indices.size() * sizeof(unsigned int)), GL_UNSIGNED_INT, 0);
 	}
 	
 private:
